requirements.txt
appdirs==1.4.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pycrypto==2.6.1
pyparsing==2.1.10
simple-crypt==4.1.7
six==1.10.0
